NBBRU (not big but relevant update)
Slightly changed how the script checks if the server is alive.
Added the possibility to type 'q', 'quit', or 'kill' when in a multi-selection prompt to close the app
Changed how the code handles config.json. This *should* be more robust and less prone to structure change.
Added the error itself to the error messages.
Updated generic try-catches in copy() and remove_fika() to only handle I/O errors.
Updated how --ip-index works when --quick is called (this is so the script doesn't hang when trying to check if the server is running)
Added some new strings in lang.json
Updated README.md to be more in-line with current capabilities